Refined the VR UI and also added info to populated objects
The UI was changed to now make use of toggle switches to determine what elements to visualise. This required some changes in the DT code as well. When objects are now populated in the VR environment, they also contain a short text description about what the reading is and what it's context is.